Find and replace Facebook -> LinkedIn
Also can you open up the issues on this repo? I have an issue in Firefox on Manjaro Linux where when I open LinkedIn in a new tab, e.g. duplicating a tab and then going to LinkedIn via the URL bar, and the new tab is moved to the end. Which is not what I want.